fix lint

+++ frontends/rioterm/src/screen/mod.rs
@@ -2529,9 +2529,10 @@ impl Screen<'_> {
             // Force invalidation for search with full damage
             {
                 let current = self.context_manager.current_mut();
-                current.renderable_content.pending_update.set_ui_damage(
-                    rio_backend::event::TerminalDamage::Full
-                );
+                current
+                    .renderable_content
+                    .pending_update
+                    .set_ui_damage(rio_backend::event::TerminalDamage::Full);
             }
         }
 
@@ -2740,28 +2741,32 @@ impl Screen<'_> {
                 if !hint_labels.is_empty() {
                     // Collect all lines that have hint labels
                     for label in hint_labels {
-                        let line = label.position.row.0 as i32 - display_offset as i32;
+                        let line = label.position.row.0 - display_offset as i32;
                         if line >= 0 && (line as usize) < screen_lines {
-                            damaged_lines.insert(rio_backend::crosswords::LineDamage::new(
-                                line as usize,
-                                true,
-                            ));
+                            damaged_lines.insert(
+                                rio_backend::crosswords::LineDamage::new(
+                                    line as usize,
+                                    true,
+                                ),
+                            );
                         }
                     }
                 }
-                
+
                 // Also damage lines with hint matches
                 if let Some(hint_matches) = &current.renderable_content.hint_matches {
                     for hint_match in hint_matches {
-                        let start_line = hint_match.start().row.0 as i32 - display_offset as i32;
-                        let end_line = hint_match.end().row.0 as i32 - display_offset as i32;
-                        
+                        let start_line = hint_match.start().row.0 - display_offset as i32;
+                        let end_line = hint_match.end().row.0 - display_offset as i32;
+
                         for line in start_line..=end_line {
                             if line >= 0 && (line as usize) < screen_lines {
-                                damaged_lines.insert(rio_backend::crosswords::LineDamage::new(
-                                    line as usize,
-                                    true,
-                                ));
+                                damaged_lines.insert(
+                                    rio_backend::crosswords::LineDamage::new(
+                                        line as usize,
+                                        true,
+                                    ),
+                                );
                             }
                         }
                     }
